pyproject.toml
[project]
name = "directory-query-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "langchain>=1.0.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4",
    "langchain-google-community>=3.0.0",
    "langchain-ollama>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "llama-cloud-services>=0.6.76",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "pymongo>=4.15.3",
    "python-dotenv>=1.1.1",
    "ruff>=0.14.1",
]
